Add AggregationQueryTest
